gossip: introduce a data collector class
We want the code to be testable. In order to achieve that we must
start extracting objects with few dependencies out of the
codebase. The first of which is the Collector class which is
responsible for managing the measurement loop and data collection.

Currently, it does everything else in the application, but shortly we
will extract other objects to take care of other relevant bits of the
application.

Signed-off-by: Felipe Balbi <felipe@balbi.sh>